requirements.txt
aerich==0.6.3
aiofiles==0.8.0
aiosqlite==0.17.0
asyncpg==0.26.0
click==8.1.3
dictdiffer==0.9.0
httptools==0.4.0
importlib-metadata==4.12.0
iso8601==1.0.2
multidict==6.0.2
pydantic==1.10.1
PyJWT==1.7.1
pypika-tortoise==0.1.6
pytz==2022.2.1
sanic==22.6.2
sanic-routing==22.3.0
tomlkit==0.11.4
tortoise-orm==0.19.2
typing_extensions==4.3.0
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
zipp==3.8.1
